Refactoring of the processor
